configuration_scripts/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0
